Commit Grain Distributions Directly
Grain distributions should be committed directly to avoid merge
conflicts and ensure reliably-timed distributions for the community.

The action will utilize a configured `ADMIN_TOKEN` secret in the github
repo to ensure the action has permissions to commit directly to master
on scheduled runs. The user of the token must have admin permissions in
the repo and the token utilize must have permissions to push to public
repositories.

test plan:
ensured committing to a protected branch worked in a forked
topocount/cred repo using a simple autocommit action